% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_last.R
\name{first_last}
\alias{first_last}
\title{Get first (minimum) and last (maximum) detection and stations of a tagged animal}
\usage{
first_last(detdf, tagidcol = "TagID", datetimecol = "DateTimeUTC",
  stationcol = "Station")
}
\arguments{
\item{detdf}{A detections dataframe}

\item{tagidcol}{Name of the column (in quotes) corresponding to individual tag IDs}

\item{datetimecol}{Name of the column corresponding to detection date and time (must be in POSIXct ymd_hm/s format)}

\item{stationcol}{Name of the column corresponding to location or station information}
}
\value{
a dataframe with five columns: TagID, first_det (first detection of tag), first_stn (first station), last_det (last detection time at final station), last_stn (station where last arrival was recorded).
}
\description{
Get first (minimum) and last (maximum) detection and stations of a tagged animal
}
\author{
Myfanwy Johnston
}
