% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_first.R
\name{last_first}
\alias{last_first}
\title{Find a tag's first detection at its last location}
\usage{
last_first(detdf, tagidcol = "TagID", datetimecol = "DateTimeUTC",
  stationcol = "Station")
}
\arguments{
\item{detdf}{a dataframe of detections for multiple tagged animals}

\item{tagidcol}{Name of the column (in quotes) corresponding to individual tag IDs}

\item{datetimecol}{Name of the column corresponding to detection date and time (must be in POSIXct ymd_hm/s format)}

\item{stationcol}{Name of the column corresponding to location or station information}
}
\value{
a dataframe with five columns: TagID, first_det (first detection of tag), first_stn (first station), last_arrival (first detection time at final station), last_stn (station where last arrival was recorded).
}
\description{
last_first: extract the first detection of a fish at its last known location.
}
\details{
This function takes a dataframe of detections and extracts the first detection of each tag, and the first detection at each tag's last station, and returns a dataframe with this information so that the travel time can then be easily calculated
}
\author{
Myfanwy Johnston
}
