% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_bardtags.R
\name{prep_bardtags}
\alias{prep_bardtags}
\title{format TagIDs for BARD query}
\usage{
prep_bardtags(
  tagids,
  df = alltags,
  tagVar = "TagID",
  codeVar = "CodeSpace",
  freq = "A69"
)
}
\arguments{
\item{tagids}{a vector of numeric TagIDs}

\item{df}{a data frame that contains (at minimum) columns for TagID and Codespace}

\item{tagVar}{The name of the TagID column in df}

\item{codeVar}{The name of the CodeSpace column in df}

\item{freq}{Frequency of tagids; defaults to "A69"}
}
\value{
A vector of tagids formatted for a query of the BARD database
}
\description{
format TagIDs for BARD query
}
\examples{
bard_tags(tagids = c(2841, 2842))

}
