% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagtales-package.R
\docType{package}
\name{tagtales-package}
\alias{tagtales-package}
\alias{tagtales}
\title{The 'tagtales' package}
\description{
This package provides functions for summarizing
    common formats of acoustic and Passive Integrated
    Transponder(PIT) tags.  It is geared toward fish acoustic
    telemetry data, and specifically data from Vemco equipment,
    but strives to be equipment-agnostic.  Summary functions
    include getting the first and last detection, generating a
    track for a tagged animal, and visualizing telemetry
    summaries.
}
